Added files from previous folder
FIles contain a headers, src, and export directory.

headers: The file headers that the user will include in his project when using the library.
src: The source files that will be private.
export: The exported library that the user will receive containing header files as well as the shared library.